0.9.6 Update Study button fix
AJAX request from clicking update study on admin page was resulting in
500 error.  Fluency rate was being passed as double (like it is in the
db) but the controller and data accessor both expected and int
(resulting in a null value).  Changed to type to double and worked as
expected.

+++ TestApp/Controllers/AjaxController.cs
@@ -87,7 +87,7 @@ namespace TestApp.Controllers
         }
 
         [HttpPost]
-        public string studyUpdate(int studyID, string hearIn, string seeIn, int hours, int minutes, int seconds, int trials, int fluency)
+        public string studyUpdate(int studyID, string hearIn, string seeIn, int hours, int minutes, int seconds, int trials, double fluency)
         {
             DataAccessor.updateStudy(studyID, hearIn, seeIn, hours, minutes, seconds, trials, fluency);
             return "Update successful.";
